+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 * Require tracer configuration instead of falling back to
   defaults, reducing logging noise.
+* The `debugTrace` tracer configuration flag has been removed in favor
+  of the `io.l5d.tracelog` telemeter.
 
 ## 0.7.5